COMP: Fix signed unsigned mismatch take two

+++ VTK/Rendering/vtkInteractorStyleRubberBandPick.cxx
@@ -141,7 +141,7 @@ void vtkInteractorStyleRubberBandPick::OnMouseMove()
   this->EndPosition[0] = this->Interactor->GetEventPosition()[0];
   this->EndPosition[1] = this->Interactor->GetEventPosition()[1];  
   int *size = this->Interactor->GetRenderWindow()->GetSize();  
-  if ((unsigned int)this->EndPosition[0] > (size[0]-1))
+  if (this->EndPosition[0] > (size[0]-1))
     {
     this->EndPosition[0] = size[0]-1;
     }
@@ -149,7 +149,7 @@ void vtkInteractorStyleRubberBandPick::OnMouseMove()
     {
     this->EndPosition[0] = 0;
     }
-  if ((unsigned int)this->EndPosition[1] > (size[1]-1))
+  if (this->EndPosition[1] > (size[1]-1))
     {
     this->EndPosition[1] = size[1]-1;
     }
